moving the saleshistory class to company folder
