+++ src/main/java/com/ticketing/system/Presentation/views/company/CompanyEventListView.java
@@ -18,7 +18,6 @@ import com.ticketing.system.Presentation.presenters.company.CompanyEventListPres
 import com.ticketing.system.Presentation.security.Capability;
 import com.ticketing.system.Presentation.security.RequireCapability;
 import com.ticketing.system.Presentation.session.AuthSession;
-import com.ticketing.system.Presentation.views.admin.CompanySalesView;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.html.Span;
